Ecore_X: Fix array problem for 64-bit (unsigned long vs unsigned int). Thanks vtorri ;)
SVN revision: 65076

+++ legacy/ecore/src/lib/ecore_x/xlib/ecore_x_window_prop.c
@@ -131,7 +131,7 @@ ecore_x_window_prop_card32_list_get(Ecore_X_Window win,
              return -1;
           }
         for (i = 0; i < num_ret; i++)
-          val[i] = ((unsigned long *)prop_ret)[i];
+          val[i] = ((unsigned int *)prop_ret)[i];
         num = num_ret;
         *plst = val;
      }
